Switches back to the production endpoint
TL;DR
-----

Returns to using the Let's Encrypt production endpoint

Details
-------

For a short time the module was hard-coded to use the Let's
Encrypt staging endpoint after running into issues with rate
limiting. This changes brings it back to using production.

To Do
-----

Set a flag or use a variable to enable either endpoint.