Shifts responsibility for k8s to the caller
TL;DR
-----

Doesn't create Kubernetes secrets within the module

Details
-------

Updates the module to not interact so it doesn't need to interact
with a Kubernetes cluster to create secrets. This both makes it
more readily reusable and allows it to be used of infrastructure
that is laid down in order to deploy Kubernetes.