Fix the missing private key output
TL;DR
-----

Fix the module to output the private key

Details
-------

The ingress for Vault wasn't getting an IP address, and it
turned out be because the certificate secret was missing the
private key. This was an artifact of switching the module to
allow a certificate request to be passed in

In order to simplify the logic for dealing with a certificate
request being passed in, I changed the code for the module to
always use a request when creating the certificate. This meant
that when a CSR is not provided, the module now creates a
private key explicitly and then creates a CSR using that key.
The output `private_key` was referring to the certificate
object to get the key, but that's only set when the CSR isn't
provided.

The fix was to output the value of the generated private key
when creating a certificate without a pre-existing CSR.