Updates dependencies for ACME module (#2)
TL;DR
-----

Finds the ACME provider where Hashicorp relocated it

Details
-------

Since Terraform now supports a heirarchical namespace of providers,
they appear to have moved the "acme" provider ourside of their
namespace and back to the `vancluever` namespace. That change
broke this module.

This change addresses that change by refererencing the provider in
it's new namespaces as a required provider for module.