Doesn't call deprecate `list()`
TL;DR
-----

Updates the use of `list()` (which is removed) to `tolist()`

Details
-------

I'd been using `list()` to create a list from a string in spite of
the deprecation warnings because I liked the way my code read when
I did. It caught up with me, because that capability was removed.

This change fixes the bug that created, and uses the `tolist()`
function instead.